Renamed PresetsWidget files to match the class name
