+++ qtEditor/PresetsWidget.h
